pyproject.toml
[tool.poetry]
name = "nebula-ai"
version = "0.1.0"
description = ""
authors = ["Nebula <me@itsnebula.net>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
aiohttp = "^3.10.10"
fast-colorthief = { git = "https://github.com/bedapisl/fast-colorthief" }
python-dotenv = "^1.0.1"
duckduckgo-search = "^6.3.0"
openai = "^1.51.2"
yt-dlp = "^2024.10.7"
faster-whisper = "^1.0.3"
beautifulsoup4 = "^4.12.3"
llvmlite = "^0.43.0"                                                      # Required for openai-whisper
numba = "^0.60.0"                                                         # Required for openai-whisper
openai-whisper = "^20240930"
aiocache = "^0.12.3"
aiohttp-socks = "^0.9.0"
asyncpraw = "^7.7.1"
asyncssh = "^2.17.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
